refactor(opentelemetry): correct Next.js span naming from the next instrumentation
Next-specific detection and naming lived in the OTel bridge (`NextSpanProcessor`,
installed by `TracerProvider`), keeping Next knowledge in the `dd-trace` package.
The bridge now exposes a generic, dependency-free pre-finish hook that the `next`
instrumentation registers the correction on. The instrumentation loads even under
`plugins: false` — the OTel-bridge setup this targets — because instrumentation
require hooks are installed independently of plugin enablement, so the hook runs
while the DD span is still unfinished and the corrected operation name / resource
reach the exported trace.
