+++ packages/dd-trace/test/opentelemetry/span_processor.spec.js
@@ -16,7 +16,6 @@ class TestSpanProcessor extends NoopSpanProcessor {
 
     this.forceFlush = sinon.stub().resolves()
     this.onStart = sinon.stub()
-    this.onEnding = sinon.stub()
     this.onEnd = sinon.stub()
     this.shutdown = sinon.stub().resolves()
   }
@@ -35,31 +34,6 @@ describe('OTel MultiSpanProcessor', () => {
     for (const processor of processors) {
       sinon.assert.calledWith(processor.onStart, 1, 2)
     }
-  })
-
-  it('should call onEnding', () => {
-    const processors = [
-      new TestSpanProcessor(),
-      new TestSpanProcessor(),
-    ]
-
-    const processor = new MultiSpanProcessor(processors)
-    processor.onEnding(3)
-
-    for (const processor of processors) {
-      sinon.assert.calledWith(processor.onEnding, 3)
-    }
-  })
-
-  it('should skip a child processor that does not implement onEnding', () => {
-    // `onEnding` is an experimental OTel hook; a user-registered processor need not implement it.
-    const withHook = new TestSpanProcessor()
-    const withoutHook = { onStart () {}, onEnd () {}, forceFlush () {}, shutdown () {} }
-
-    const processor = new MultiSpanProcessor([withoutHook, withHook])
-    processor.onEnding(3)
-
-    sinon.assert.calledWith(withHook.onEnding, 3)
   })
 
   it('should call onEnd', () => {
